requirements.txt
Flask
pandas
scikit-learn
nltk
pyspellchecker